+++ src/ui.h
@@ -517,6 +517,7 @@ public:
 
     static void ScreenChangeFixExportColors(int link, uint32_t v);
     static void ScreenChangeBackFaces(int link, uint32_t v);
+    static void ScreenChangeShowContourAreas(int link, uint32_t v);
     static void ScreenChangeCheckClosedContour(int link, uint32_t v);
     static void ScreenChangePwlCurves(int link, uint32_t v);
     static void ScreenChangeCanvasSizeAuto(int link, uint32_t v);